Grammatical Change
changed  the last question 
from 

Create an nginx deployment of 2 replicas, expose it via a ClusterIP service on port 80. Create a NetworkPolicy so that only pods with labels 'access: granted' can access the deployment and apply it

to

Create an nginx deployment of 2 replicas, expose it via a ClusterIP service on port 80. Create a NetworkPolicy so that only pods with labels 'access: granted' can access the pods created by nginx deployment and apply it

AFAIK Pods can access other Pods and Pods cannot access Deployments. So I got confused when I read the question for the first time.